backend/requirements.txt
annotated-types
anyio
certifi
cffi
charset-normalizer
cryptography
Deprecated
distro
h11
httpcore
httpx
idna
openai
pycparser
pydantic
pydantic_core
PyGithub
PyJWT
PyNaCl
requests
setuptools
sniffio
tqdm
typing_extensions
urllib3
wheel
wrapt
langchain
langchain_openai
langchain_community
fastapi[standard]
httpx
